+++ R/build-home.R
@@ -121,14 +121,7 @@
 #'     href: http://website.com
 #' ```
 #'
-#' You can completely remove the sidebar.
-#'
-#' ```
-#' home:
-#'   sidebar: FALSE
-#' ```
-#'
-#' You can provide a ready-made sidebar HTML.
+#' You can provide a ready-made sidebar HTML:
 #'
 #' ```
 #' home:
@@ -136,8 +129,8 @@
 #'     html: path-to-sidebar.html
 #' ```
 #'
-#' You can change the order of sidebar elements (except Dev Status):
-#' links, license (with an "s"), community, citation, authors
+#' You can change the order of sidebar elements:
+#' links, license (with an "s"), community, citation, authors, dev (badges)
 #' and add custom ones.
 #' The example below creates a sidebar whose only elements will be the
 #' authors section, a custom section, and a Dev Status section if there are
@@ -148,11 +141,18 @@
 #' ```
 #' home:
 #'   sidebar:
-#'     structure: [authors, custom]
+#'     structure: [authors, custom, dev]
 #'     components:
 #'       custom:
 #'         title: Funding
 #'         text: We are grateful for funding!
+#' ```
+#'
+#' You can completely remove the sidebar.
+#'
+#' ```
+#' home:
+#'   sidebar: FALSE
 #' ```
 #'
 #' READMEs usually start with an `<h1>` containing the package name. If