+++ NEWS.md
@@ -1,5 +1,12 @@
 # pkgdown (development version)
 
+* Make sidebar specification more flexible: users can now
+    * change the order of sidebar elements
+    * add custom sidebar sections (title, text that has to be HTML)
+    * completely suppress the navbar (even "Dev status")
+    * provide their own HTML for the navbar. (#1443, #1488)
+
+
 * Protect the rules drawn by the CLI (as for example, in `build_site()`) against
   very narrow terminal windows with small `getOption('width')`s 
   (@maxheld83, #1435).